catch and wrap fatal exceptions to allow them to be thrown from future
by default, Future only reports NonFatal exceptions as errors. Fatals are
swallowed and propagated to the running thread, which silently kills the thread
because the exceution context does not attach uncaught exception handlers to the
created threads. It might be better to use an EC that attaches UCE handlers later.